ENGR00224997-2 vold: refine the return value check of fat fsck
When the volume checking each partitions for avaliable
VFAT filesystem, we will get ENOEXEC return value on
exec msdos_fsck who give out "Invalid cluster size: 0"
on the extension partition.
This error return, should not block volume manager
to scan the left partitions as it's not a I/O error.
This issue happend when we have a extension partition
on external SD card.

Change-Id: I5f9e269c6d07fa58ef19a4be2e216ece1a9d12e0
Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>